adventures/15_cloud_run_vertex_proxy/app/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
google-cloud-aiplatform==1.49.0
pydantic==2.7.1
